feat: set storage object on dataset open
