+++ test/e2e/storage-open-return-storage-object/actor/.actor/actor.json
